Intel back tests + all size_t changed unsigned int

+++ include/tools/printing.hpp
@@ -8,17 +8,17 @@
 using namespace std;
 #include "hint.hpp"
 
-template <typename val, size_t W, bool is_signed, template<size_t, bool> class wrapper>
+template <typename val, unsigned int W, bool is_signed, template<unsigned int, bool> class wrapper>
 struct printer{
     inline constexpr void operator()(stringstream & s, hint_base<W, is_signed, wrapper> const & signal)
     {
         s << (signal.template isSet<val::value>() ? '1' : '0');
-        printer<integral_constant<size_t, val::value-1>, W, is_signed, wrapper>{}(s, signal);
+        printer<integral_constant<unsigned int, val::value-1>, W, is_signed, wrapper>{}(s, signal);
     }
 };
 
-template <size_t W, bool is_signed, template<size_t, bool> class wrapper>
-struct printer<integral_constant<size_t, 0>, W, is_signed, wrapper>
+template <unsigned int W, bool is_signed, template<unsigned int, bool> class wrapper>
+struct printer<integral_constant<unsigned int, 0>, W, is_signed, wrapper>
 {
     inline constexpr void operator()(stringstream& s, hint_base<W, is_signed, wrapper> const & signal)
     {
@@ -27,11 +27,11 @@ struct printer<integral_constant<size_t, 0>, W, is_signed, wrapper>
 };
 
 
-template <size_t W, bool is_signed, template<size_t, bool> class wrapper>
+template <unsigned int W, bool is_signed, template<unsigned int, bool> class wrapper>
 string to_string(hint_base<W, is_signed, wrapper> const & signal)
 {
     stringstream s{};
-    printer<integral_constant<size_t, W-1>, W, is_signed, wrapper>{}(s, signal);
+    printer<integral_constant<unsigned int, W-1>, W, is_signed, wrapper>{}(s, signal);
     return s.str();
 }
